Added Death Date of Dorothea Kolb.
Namechange again due to Windows' stupid "This file is already opened in another process" without an option to kill this other process or to "save it anyways". I miss AmigaOS, which did not try to be more clever than the user! If I invalidate the data of a different process it is *MY DECISION*, not something the OS should block !!!!