Add files via upload
Added possibility to use the Converter for Death Dates. It is now at the same stage like the old Excel Addin which then is obsolete.
Note that the Deaths Module does not support some things (like Burial Date) as in the Records I was using it for no Burial Date was given. Eventually it has to be expanded to support this. But not now. The old Excel Addin had the same flaw.